9C-280 - Includes the sort logic in the drawer

+++ modules/process/src/main/scala/com/fortysevendeg/ninecardslauncher/process/device/models/Models.scala
@@ -10,8 +10,8 @@ case class App(
   category: String,
   imagePath: String,
   colorPrimary: String,
-  dateInstalled: Double,
-  dateUpdate: Double,
+  dateInstalled: Long,
+  dateUpdate: Long,
   version: String,
   installedFromGooglePlay: Boolean)
